requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorzero==2.0
commonmark==0.9.1
cryptography==42.0.1
decorator==5.1.1
executing==2.0.1
fastapi==0.109.0
gpiozero==2.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
jedi==0.19.1
Jinja2==3.1.3
littleutils==0.2.2
mariadb==1.1.9
MarkupSafe==2.1.3
packaging==23.2
paramiko==3.4.0
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
platformdirs==2.6.2
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyNaCl==1.5.0
pyscript==0.2.5
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rich==12.6.0
rich-click==1.7.3
RPi.GPIO==0.7.1
six==1.16.0
sniffio==1.3.0
sorcery==0.2.2
SQLAlchemy==2.0.25
sqlmodel==0.0.14
stack-data==0.6.3
starlette==0.35.1
toml==0.10.2
tornado==6.4
traitlets==5.14.1
typer==0.4.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
wcwidth==0.2.13
webssh==1.6.2
wrapt==1.16.0
